[MEAR-292] - Don't add MANFIEST.mf Class-Path entry for EAR module if skipClassPathModification is true and original MANFIEST.mf of EAR module artifact doesn't contain Class-Path entry.

+++ src/main/java/org/apache/maven/plugins/ear/EarMojo.java
@@ -816,12 +816,15 @@ public class EarMojo
             Attribute classPath = mf.getMainSection().getAttribute( "Class-Path" );
             List<String> classPathElements = new ArrayList<String>();
 
+            boolean classPathExisted;
             if ( classPath != null )
             {
+                classPathExisted = true;
                 classPathElements.addAll( Arrays.asList( classPath.getValue().split( " " ) ) );
             }
             else
             {
+                classPathExisted = false;
                 classPath = new Attribute( "Class-Path", "" );
             }
 
@@ -888,6 +891,7 @@ public class EarMojo
             }
 
             // Modify the classpath entries in the manifest
+            boolean forceClassPathModification = javaEEVersion.lt( JavaEEVersion.FIVE ) || defaultLibBundleDir == null;
             for ( EarModule o : getModules() )
             {
                 if ( o instanceof JarModule )
@@ -898,30 +902,27 @@ public class EarMojo
                     {
                         classPathElements.set( moduleClassPathIndex, jm.getUri() );
                     }
-                    else
+                    else if ( !skipClassPathModification )
                     {
-                        if ( !skipClassPathModification )
-                        {
-                            classPathElements.add( jm.getUri() );
-                        }
-                        else
-                        {
-                            if ( javaEEVersion.lt( JavaEEVersion.FIVE ) || defaultLibBundleDir == null )
-                            {
-                                classPathElements.add( jm.getUri() );
-                            }
-                        }
+                        classPathElements.add( jm.getUri() );
+                    }
+                    else if ( forceClassPathModification )
+                    {
+                        classPathElements.add( jm.getUri() );
                     }
                 }
             }
-            classPath.setValue( StringUtils.join( classPathElements.iterator(), " " ) );
-            mf.getMainSection().addConfiguredAttribute( classPath );
-
-            // Write the manifest to disk
-            try ( FileOutputStream out = new FileOutputStream( manifestFile );
-                  OutputStreamWriter writer = new OutputStreamWriter( out, StandardCharsets.UTF_8 ) )
+            if ( !skipClassPathModification || !classPathElements.isEmpty() || classPathExisted )
             {
-                mf.write( writer );
+                classPath.setValue( StringUtils.join( classPathElements.iterator(), " " ) );
+                mf.getMainSection().addConfiguredAttribute( classPath );
+
+                // Write the manifest to disk
+                try ( FileOutputStream out = new FileOutputStream( manifestFile );
+                      OutputStreamWriter writer = new OutputStreamWriter( out, StandardCharsets.UTF_8 ) )
+                {
+                    mf.write( writer );
+                }
             }
 
             if ( original.isFile() )
